requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.1

# LangChain and LLM - updated to use OpenAI
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# Search tool
tavily-python>=0.3.0

# Web scraping dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
langchain-community>=0.2.0

# Data processing - updated for Python 3.13 compatibility
pandas>=2.2.0

# HTTP client for examples
requests>=2.32.0
aiohttp>=3.9.1 